# Performance tests and fix for IPAM controller.