Tests the four modes of allocations. Can be run using
./test-performance.sh under tests/integration/ipamperf
directory. See ./test-performance.sh -h for supported flags.

<!--  Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, read our contributor guidelines https://git.k8s.io/community/contributors/devel/pull-requests.md#the-pr-submit-process and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/devel/pull-requests.md#best-practices-for-faster-reviews
3. Follow the instructions for writing a release note: https://git.k8s.io/community/contributors/devel/pull-requests.md#write-release-notes-if-needed
4. If the PR is unfinished, see how to mark it: https://github.com/kubernetes/community/blob/master/contributors/devel/pull-requests.md#marking-unfinished-pull-requests
-->

**What this PR does / why we need it**:
Builds a test infrastructure for performance validation of IPAM controllers.

**Which issue(s) this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close the issue(s) when PR gets merged)*:
Fixes #

**Special notes for your reviewer**:
The results are logged as JSON. Here is a sample from one of the runs.
```json
[
  {
    "Name": "TestPerformance/IPAMFromClusterAllocatorKubeQPS10Nodes10",
    "Config": {
      "CreateQPS": 10,
      "KubeQPS": 10,
      "CloudQPS": 10,
      "NumNodes": 10,
      "AllocatorType": "IPAMFromCluster",
      "Cloud": {
        "ClusterID": {},
        "AlphaFeatureGate": null
      }
    },
    "Succeeded": true,
    "MaxAllocTime": "967.172379ms",
    "TotalAllocTime": "992.017781ms",
    "NodeAllocTime": [
      {
        "Name": "sample-node-pkcz4",
        "PodCIDR": "10.96.9.0/24",
        "Duration": "967.172379ms"
      },
      {
        "Name": "sample-node-nt858",
        "PodCIDR": "10.96.8.0/24",
        "Duration": "869.112898ms"
      },
      {
        "Name": "sample-node-782ln",
        "PodCIDR": "10.96.7.0/24",
        "Duration": "772.920979ms"
      },
      {
        "Name": "sample-node-d99rg",
        "PodCIDR": "10.96.6.0/24",
        "Duration": "674.003487ms"
      },
      {
        "Name": "sample-node-5lf5b",
        "PodCIDR": "10.96.5.0/24",
        "Duration": "477.069302ms"
      },
      {
        "Name": "sample-node-c8jf8",
        "PodCIDR": "10.96.0.0/24",
        "Duration": "10.605318ms"
      },
      {
        "Name": "sample-node-t8jt6",
        "PodCIDR": "10.96.1.0/24",
        "Duration": "8.948069ms"
      },
      {
        "Name": "sample-node-tpdmw",
        "PodCIDR": "10.96.3.0/24",
        "Duration": "8.213681ms"
      },
      {
        "Name": "sample-node-clkl8",
        "PodCIDR": "10.96.4.0/24",
        "Duration": "7.465896ms"
      },
      {
        "Name": "sample-node-mmxl6",
        "PodCIDR": "10.96.2.0/24",
        "Duration": "7.184109ms"
      }
    ]
  }
]
```

**Release note**:
<!--  Write your release note:
1. Enter your extended release note in the below block. If the PR requires additional action from users switching to the new release, include the string "action required".
2. If no release note is required, just write "NONE".
-->
```release-note
Integration tests for performance validation of IPAM allocators
```

/cc @bowei , @jingax10 , @shyamjvs