@shyamjvs 2018-03-08
I've done an initial pass, and my feeling is that large parts of this PR can probably be eliminated by trying to re-use existing code. If you need some references, I'm happy to help.