@satyasm 2018-03-08
Clarified why most of the code exists. The idea was to mock the edges (namely server) and run most of the client code as is...