@shyamjvs 2018-03-08
SG. I did a second pass and added bunch of comments.
Overall it looks in good shape, will hopefully try to finish reviewing by tomorrow.