@satyasm 2018-03-12
fixed most of the comments, clarified a few. working on replacing watch with informer.