@shyamjvs 2018-03-12
Couple of minor comments. Will lgtm once they're addressed.
Also, due to severe lack of free cycles, I'm going to cut the review down at this point :)
As we probably can anyway iterate on the code later.