@shyamjvs 2018-03-12
Ping @bowei - For approving the pkg/controller/nodeipam part.