@bowei 2018-03-13
Can you separate the bug fixes into a separate commit?

Also, I don't see any description of what the bug fix did in the commit message.