@satyasm 2018-03-13
@bowei this is the baseline code, on which further changes like introducing the delay etc is done. The idea here as I also explained to shyam was to run the whole of the GCECloud client code as is and introduce a mock server that meets expectations. Once we have this baseline, we will be able to add things like delay at the server, track server side QPS, add behavior for server limits etc.,. Hoping we can get this baseline code in and then make smaller PR's for the changes.