@satyasm 2018-03-13
can also separate the sync.go code changes from the test code if that is easier.