@satyasm 2018-03-13
removed the ipam/controller.go changes and made this PR pure test code. Let's keep this open for now. Will do a different PR by mocking the cloud client and we can compare and pick the one we want. Will do a separate PR for the controller.go changes.